move data loading to .ts file
