+++ frontend/src/routes/+layout.server.ts
@@ -2,7 +2,6 @@ import type { LayoutServerLoad } from "./$types";
 
 export const load: LayoutServerLoad = async ({ locals }) => {
   return {
-    user: locals.user,
-    menu: locals.menu
+    user: locals.user
   };
 };
